@@ -1,8 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <OutputType>Exe</OutputType>
-        <TargetFramework>netcoreapp3.1</TargetFramework>
+        <TargetFramework>net6.0</TargetFramework>
     </PropertyGroup>
 
     <ItemGroup>
@@ -18,6 +17,7 @@
       <Compile Remove="NeuralNetwork\Networks\Perceptron.cs" />
       <Compile Remove="NeuralNetwork\Networks\**" />
       <Compile Remove="NeuralNetwork\IInputNeuron.cs" />
+      <Compile Remove="Program.cs" />
     </ItemGroup>
 
     <ItemGroup>
